// src/app/auth/models/user.ts
export class User {
  id: number;
  email: string;
  password: string;
  name:string;
  lastName:string;
  given_name: string;
  family_name: string;
  phone: number;
  //avatar: string;
  role: string;
  token?: string;
  cin: number;
  typecontrat: string;
  salaire: number;
  salairebrut: number;
  permission: number;
  photo: string;
}